+++ cfdm/cfdmimplementation.py
@@ -1106,7 +1106,7 @@ class CFDMImplementation(Implementation):
     def nc_get_dataset_shards(self, data):
         """Get the Zarr dataset sharding strategy for the data.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.0.0
 
         .. seealso:: `nc_set_dataset_shards`
 
@@ -1267,7 +1267,7 @@ class CFDMImplementation(Implementation):
     def nc_set_dataset_shards(self, data, shards):
         """Set the Zarr dataset sharding strategy for the data.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.0.0
 
         .. seealso:: `nc_get_dataset_shards`