+++ cfdm/mixin/propertiesdata.py
@@ -542,7 +542,7 @@ class PropertiesData(Properties):
 
             {{data: `bool` or `None`, optional}}
 
-                .. versionadded:: (cfdm) NEXTVERSION
+                .. versionadded:: (cfdm) 1.13.0.0
 
             display: `bool`, optional
                 If False then return the description as a string. By